controllers created

+++ src/main/java/com/etiya/ecommerceDemo/repositories/abstracts/ReviewReplyDao.java
@@ -3,5 +3,5 @@ package com.etiya.ecommerceDemo.repositories.abstracts;
 import com.etiya.ecommerceDemo.entities.concretes.ReviewReply;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReviewReplyDao extends JpaRepository<ReviewReply,Integer> {
+public interface ReviewReplyDao extends JpaRepository<ReviewReply,Long> {
 }
